encrypted-traffic-classifier-flask/requirements.txt
Flask==2.0.3
Flask_Cors==4.0.1
loguru==0.7.2
Pillow==10.4.0
timm==1.0.10
torch==1.11.0
torchvision==0.12.0
Werkzeug==2.1.2